// src/chessBoard/objects/chessColors.d.ts
interface ChessColorData {
    selectId: string;
    default: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    woodland: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    checker: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    sky: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    funKingdom: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    classic: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    viking: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
    fireAndIce: {
        primary: string;
        secondary: string;
        selectValue: string;
    };
}
export declare const ChessColorData: ChessColorData;
export {};